IGAPP-1256: remove placeholder icons web
